Fix all of the tests
To avoid circular dependencies, the tests had to be moved
to a separate cargo project within the workspace.
