+++ av_metrics/src/lib.rs
@@ -43,14 +43,3 @@ pub enum MetricsError {
     #[error("Unreachable")]
     NonExhaustive,
 }
-
-#[cfg(test)]
-#[inline(always)]
-fn assert_metric_eq(expected: f64, value: f64) {
-    assert!(
-        (expected - value).abs() < 0.01,
-        "Expected {}, got {}",
-        expected,
-        value
-    );
-}
