+++ av_metrics/src/video/ciede/mod.rs
@@ -548,91 +548,3 @@ mod avx2 {
     impl DeltaEAVX2 for BD10 {}
     impl DeltaEAVX2 for BD12 {}
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::assert_metric_eq;
-    use std::fs::File;
-    use y4m::Decoder;
-
-    #[test]
-    fn ciede2000_yuv420p8_nosimd() {
-        let mut file1 = File::open("./testfiles/yuv420p8_input.y4m").unwrap();
-        let mut dec1 = Decoder::new(&mut file1).unwrap();
-        let mut file2 = File::open("./testfiles/yuv420p8_output.y4m").unwrap();
-        let mut dec2 = Decoder::new(&mut file2).unwrap();
-        let result = calculate_video_ciede_nosimd::<_>(&mut dec1, &mut dec2, None).unwrap();
-        assert_metric_eq(36.2821, result);
-    }
-
-    #[test]
-    fn ciede2000_yuv420p8() {
-        let mut file1 = File::open("./testfiles/yuv420p8_input.y4m").unwrap();
-        let mut dec1 = Decoder::new(&mut file1).unwrap();
-        let mut file2 = File::open("./testfiles/yuv420p8_output.y4m").unwrap();
-        let mut dec2 = Decoder::new(&mut file2).unwrap();
-        let result = calculate_video_ciede::<_>(&mut dec1, &mut dec2, None).unwrap();
-        assert_metric_eq(36.2821, result);
-    }
-
-    #[test]
-    fn ciede2000_yuv422p8_nosimd() {
-        let mut file1 = File::open("./testfiles/yuv422p8_input.y4m").unwrap();
-        let mut dec1 = Decoder::new(&mut file1).unwrap();
-        let mut file2 = File::open("./testfiles/yuv422p8_output.y4m").unwrap();
-        let mut dec2 = Decoder::new(&mut file2).unwrap();
-        let result = calculate_video_ciede_nosimd::<_>(&mut dec1, &mut dec2, None).unwrap();
-        assert_metric_eq(43.9618, result);
-    }
-
-    #[test]
-    fn ciede2000_yuv422p8() {
-        let mut file1 = File::open("./testfiles/yuv422p8_input.y4m").unwrap();
-        let mut dec1 = Decoder::new(&mut file1).unwrap();
-        let mut file2 = File::open("./testfiles/yuv422p8_output.y4m").unwrap();
-        let mut dec2 = Decoder::new(&mut file2).unwrap();
-        let result = calculate_video_ciede::<_>(&mut dec1, &mut dec2, None).unwrap();
-        assert_metric_eq(43.9618, result);
-    }
-
-    #[test]
-    fn ciede2000_yuv444p8_nosimd() {
-        let mut file1 = File::open("./testfiles/yuv444p8_input.y4m").unwrap();
-        let mut dec1 = Decoder::new(&mut file1).unwrap();
-        let mut file2 = File::open("./testfiles/yuv444p8_output.y4m").unwrap();
-        let mut dec2 = Decoder::new(&mut file2).unwrap();
-        let result = calculate_video_ciede_nosimd::<_>(&mut dec1, &mut dec2, None).unwrap();
-        assert_metric_eq(37.5106, result);
-    }
-
-    #[test]
-    fn ciede2000_yuv444p8() {
-        let mut file1 = File::open("./testfiles/yuv444p8_input.y4m").unwrap();
-        let mut dec1 = Decoder::new(&mut file1).unwrap();
-        let mut file2 = File::open("./testfiles/yuv444p8_output.y4m").unwrap();
-        let mut dec2 = Decoder::new(&mut file2).unwrap();
-        let result = calculate_video_ciede::<_>(&mut dec1, &mut dec2, None).unwrap();
-        assert_metric_eq(37.5106, result);
-    }
-
-    #[test]
-    fn ciede2000_yuv420p10_nosimd() {
-        let mut file1 = File::open("./testfiles/yuv420p10_input.y4m").unwrap();
-        let mut dec1 = Decoder::new(&mut file1).unwrap();
-        let mut file2 = File::open("./testfiles/yuv420p10_output.y4m").unwrap();
-        let mut dec2 = Decoder::new(&mut file2).unwrap();
-        let result = calculate_video_ciede_nosimd::<_>(&mut dec1, &mut dec2, None).unwrap();
-        assert_metric_eq(36.3691, result);
-    }
-
-    #[test]
-    fn ciede2000_yuv420p10() {
-        let mut file1 = File::open("./testfiles/yuv420p10_input.y4m").unwrap();
-        let mut dec1 = Decoder::new(&mut file1).unwrap();
-        let mut file2 = File::open("./testfiles/yuv420p10_output.y4m").unwrap();
-        let mut dec2 = Decoder::new(&mut file2).unwrap();
-        let result = calculate_video_ciede::<_>(&mut dec1, &mut dec2, None).unwrap();
-        assert_metric_eq(36.3691, result);
-    }
-}
