+++ av_metrics/src/video/psnr.rs
@@ -177,115 +177,3 @@ fn calculate_plane_total_squared_error<T: Pixel>(plane1: &Plane<T>, plane2: &Pla
         .map(|err| err * err)
         .sum::<u64>() as f64
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::assert_metric_eq;
-    use std::fs::File;
-    use y4m::Decoder;
-
-    #[test]
-    fn psnr_yuv420p8() {
-        let mut file1 = File::open("./testfiles/yuv420p8_input.y4m").unwrap();
-        let mut dec1 = Decoder::new(&mut file1).unwrap();
-        let mut file2 = File::open("./testfiles/yuv420p8_output.y4m").unwrap();
-        let mut dec2 = Decoder::new(&mut file2).unwrap();
-        let result = calculate_video_psnr::<_>(&mut dec1, &mut dec2, None).unwrap();
-        assert_metric_eq(32.5281, result.y);
-        assert_metric_eq(36.4083, result.u);
-        assert_metric_eq(39.8238, result.v);
-        assert_metric_eq(33.6861, result.avg);
-    }
-
-    #[test]
-    fn psnr_yuv422p8() {
-        let mut file1 = File::open("./testfiles/yuv422p8_input.y4m").unwrap();
-        let mut dec1 = Decoder::new(&mut file1).unwrap();
-        let mut file2 = File::open("./testfiles/yuv422p8_output.y4m").unwrap();
-        let mut dec2 = Decoder::new(&mut file2).unwrap();
-        let result = calculate_video_psnr::<_>(&mut dec1, &mut dec2, None).unwrap();
-        assert_metric_eq(38.6740, result.y);
-        assert_metric_eq(47.5219, result.u);
-        assert_metric_eq(48.8615, result.v);
-        assert_metric_eq(41.2190, result.avg);
-    }
-
-    #[test]
-    fn psnr_yuv444p8() {
-        let mut file1 = File::open("./testfiles/yuv444p8_input.y4m").unwrap();
-        let mut dec1 = Decoder::new(&mut file1).unwrap();
-        let mut file2 = File::open("./testfiles/yuv444p8_output.y4m").unwrap();
-        let mut dec2 = Decoder::new(&mut file2).unwrap();
-        let result = calculate_video_psnr::<_>(&mut dec1, &mut dec2, None).unwrap();
-        assert_metric_eq(32.4235, result.y);
-        assert_metric_eq(40.1212, result.u);
-        assert_metric_eq(43.1900, result.v);
-        assert_metric_eq(36.2126, result.avg);
-    }
-
-    #[test]
-    fn psnr_yuv420p10() {
-        let mut file1 = File::open("./testfiles/yuv420p10_input.y4m").unwrap();
-        let mut dec1 = Decoder::new(&mut file1).unwrap();
-        let mut file2 = File::open("./testfiles/yuv420p10_output.y4m").unwrap();
-        let mut dec2 = Decoder::new(&mut file2).unwrap();
-        let result = calculate_video_psnr::<_>(&mut dec1, &mut dec2, None).unwrap();
-        assert_metric_eq(32.5421, result.y);
-        assert_metric_eq(36.4922, result.u);
-        assert_metric_eq(39.8558, result.v);
-        assert_metric_eq(33.7071, result.avg);
-    }
-
-    #[test]
-    fn apsnr_yuv420p8() {
-        let mut file1 = File::open("./testfiles/yuv420p8_input.y4m").unwrap();
-        let mut dec1 = Decoder::new(&mut file1).unwrap();
-        let mut file2 = File::open("./testfiles/yuv420p8_output.y4m").unwrap();
-        let mut dec2 = Decoder::new(&mut file2).unwrap();
-        let result = calculate_video_apsnr::<_>(&mut dec1, &mut dec2, None).unwrap();
-        assert_metric_eq(32.5450, result.y);
-        assert_metric_eq(36.4087, result.u);
-        assert_metric_eq(39.8244, result.v);
-        assert_metric_eq(33.6995, result.avg);
-    }
-
-    #[test]
-    fn apsnr_yuv422p8() {
-        let mut file1 = File::open("./testfiles/yuv422p8_input.y4m").unwrap();
-        let mut dec1 = Decoder::new(&mut file1).unwrap();
-        let mut file2 = File::open("./testfiles/yuv422p8_output.y4m").unwrap();
-        let mut dec2 = Decoder::new(&mut file2).unwrap();
-        let result = calculate_video_apsnr::<_>(&mut dec1, &mut dec2, None).unwrap();
-        assert_metric_eq(38.6741, result.y);
-        assert_metric_eq(47.5219, result.u);
-        assert_metric_eq(48.8616, result.v);
-        assert_metric_eq(41.2191, result.avg);
-    }
-
-    #[test]
-    fn apsnr_yuv444p8() {
-        let mut file1 = File::open("./testfiles/yuv444p8_input.y4m").unwrap();
-        let mut dec1 = Decoder::new(&mut file1).unwrap();
-        let mut file2 = File::open("./testfiles/yuv444p8_output.y4m").unwrap();
-        let mut dec2 = Decoder::new(&mut file2).unwrap();
-        let result = calculate_video_apsnr::<_>(&mut dec1, &mut dec2, None).unwrap();
-        assert_metric_eq(32.4412, result.y);
-        assert_metric_eq(40.1264, result.u);
-        assert_metric_eq(43.1943, result.v);
-        assert_metric_eq(36.2271, result.avg);
-    }
-
-    #[test]
-    fn apsnr_yuv420p10() {
-        let mut file1 = File::open("./testfiles/yuv420p10_input.y4m").unwrap();
-        let mut dec1 = Decoder::new(&mut file1).unwrap();
-        let mut file2 = File::open("./testfiles/yuv420p10_output.y4m").unwrap();
-        let mut dec2 = Decoder::new(&mut file2).unwrap();
-        let result = calculate_video_apsnr::<_>(&mut dec1, &mut dec2, None).unwrap();
-        assert_metric_eq(32.5586, result.y);
-        assert_metric_eq(36.4923, result.u);
-        assert_metric_eq(39.8563, result.v);
-        assert_metric_eq(33.7200, result.avg);
-    }
-}
